youtube_uploader.py
import os
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload
from dotenv import load_dotenv

load_dotenv()

class YouTubeUploader:
    def __init__(self, credentials_path=None):
        self.credentials_path = credentials_path or os.getenv("YOUTUBE_CREDENTIALS_PATH")
        self.youtube = None

    def authenticate(self):
        """Authenticates with the YouTube Data API."""
        # Note: This is a stub for the full OAuth2 flow.
        # In a real scenario, this would involve loading tokens or finishing a flow.
        print("INFO: YouTube Data API Authentication stubbed.")
        return True

    def upload_video(self, file_path, title, description, tags=None, category_id="27"):
        """Uploads a video to YouTube."""
        if not self.youtube:
            print("WARNING: YouTube client not initialized. Simulation upload...")
            print(f"Uploading {file_path} as '{title}'...")
            return {"status": "Simulated Success", "video_id": "SIM_12345"}
        
        # Real upload logic would go here using MediaFileUpload
        return {"status": "Success"}

if __name__ == "__main__":
    uploader = YouTubeUploader()
    # uploader.upload_video("test.mp4", "My Historical Video", "Deep dive into history.")
